# docs/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Login & User Management Microservice API
  description: Servicio de autenticación, registro y gestión de usuarios (JWT + Refresh Token).
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor Local

paths:
  /auth/login:
    post:
      summary: Autenticar usuario y obtener tokens
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'

  /auth/register:
    post:
      summary: Registrar un nuevo usuario
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Usuario creado con éxito
        '400':
          $ref: '#/components/responses/ErrorResponse'

  /auth/revoke:
    post:
      summary: Invalidar una sesión (Logout)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Token revocado

  /auth/validate:
    post:
      summary: Validar un token existente
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Token válido (devuelve boolean)
          content:
            application/json:
              schema:
                type: boolean

  /users:
    get:
      summary: Listar todos los usuarios
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'

  /users/id/{id}:
    get:
      summary: Obtener usuario por ID
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /users/dni/{dni}:
    get:
      summary: Obtener usuario por DNI
      tags: [Users]
      parameters:
        - name: dni
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  /users/search:
    get:
      summary: Buscar usuarios por nombre
      tags: [Users]
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'

  /users/{id}:
    put:
      summary: Actualizar datos de un usuario
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      security:
        - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    delete:
      summary: Eliminar un usuario
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Usuario eliminado con éxito

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: "admin@empresa.com" }
        password: { type: string, format: password, example: "Password123" }

    RegisterRequest:
      type: object
      required: [email, password, name, lastname, phone, dni, roles]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        name: { type: string }
        lastname: { type: string }
        phone: { type: string, pattern: '^[0-9]{6,10}$' }
        dni: { type: string, pattern: '^[1-9][0-9]{5,9}$' }
        roles:
          type: array
          items:
            type: string
            enum: [USER, ADMIN]

    TokenRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }

    TokenResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, description: "Segundos de validez" }

    UserUpdateRequest:
      type: object
      properties:
        name: { type: string }
        lastname: { type: string }
        phone: { type: string }
        isActive: { type: boolean }

    UserResponse:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        lastname: { type: string }
        phone: { type: string }
        dni: { type: string }
        isActive: { type: boolean }
        roles:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: string }

    ErrorResponse:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        timestamp: { type: string, format: date-time }

  responses:
    ErrorResponse:
      description: Error en la operación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'