# src/main/java/microservice/user/reactive/login/infraestructure/driver/rest/dto/resp/TokenResponse.java
package microservice.user.reactive.login.infraestructure.driver.rest.dto.resp;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class TokenResponse {
    private String accessToken;
    private String refreshToken;
    private Long expiresIn;
}